+++ tests/common.h
@@ -4,8 +4,8 @@
 #include <vector>
 #include <cstdint>
 
-void fillVectorRandomUniform(uint64_t min, uint64_t max, std::vector<unsigned int> *vector);
-void fillVectorRandomGeometric(std::vector<unsigned int> *vector);
+void fillVectorRandomUniform(uint64_t min, uint64_t max, std::vector<uint64_t> *vector);
+void fillVectorRandomGeometric(std::vector<uint64_t> *vector);
 
 
 #endif  // GABAC_TESTS_COMMON_H_
